pyproject.toml
[tool.poetry]
name = "markett"
version = "0.0.1"
description = ""
authors = ["parisa <parisafarivash@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.5"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
pytest-django = "^4.9.0"
pyotp = "^2.9.0"
twilio = "^9.2.4"

[tool.poetry.group.dev.dependencies]
djangorestframework = "^3.15.2"
Django = "^5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
